Issuse #17932: Fix an integer overflow issue on Windows 64-bit in tuple iterators: change the C type of tupleiterobject.it_index from long to Py_ssize_t.

+++ Objects/tupleobject.c
@@ -930,7 +930,7 @@ PyTuple_Fini(void)
 
 typedef struct {
     PyObject_HEAD
-    long it_index;
+    Py_ssize_t it_index;
     PyTupleObject *it_seq; /* Set to NULL when iterator is exhausted */
 } tupleiterobject;
